Moved wrapper_types and type_dispatcher to a new types/ directory.
